order is now saved to database

+++ src/main/java/com/backend/entity/OrderUserId.java
@@ -7,23 +7,23 @@ import java.util.Objects;
 @Embeddable
 public class OrderUserId implements Serializable {
 
-    private Long orderId;
+    private int orderId;
     private Long userId;
 
     // Constructors, Getters, Setters, Equals, and HashCode
 
     public OrderUserId() {}
 
-    public OrderUserId(Long orderId, Long userId) {
+    public OrderUserId(int orderId, Long userId) {
         this.orderId = orderId;
         this.userId = userId;
     }
 
-    public Long getOrderId() {
+    public int getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Long orderId) {
+    public void setOrderId(int orderId) {
         this.orderId = orderId;
     }
